BUGFIX: exchangeable was passing the wrong part of the array to the next image
Changed default_halo_size to 1 which broke code.  It turns out we were passing the halo region from one image to the next
image instead of passing the processed outer portion of the local domain to the halo of the next image.

Also updated halo code to use the upper and lower bounds of the array instead of 1:size